+++ frontend/src/employee-frontend/components/person-profile/PersonDependants.tsx
@@ -20,7 +20,7 @@ import { useTranslation } from '../../state/i18n'
 import { formatName } from '../../utils'
 import { renderResult } from '../async-rendering'
 
-import { NameTd } from './PersonProfile'
+import { NameTd } from './common'
 
 const getPersonDependantsResult = wrapResult(getPersonDependants)
